Refactor UserProfile components to handle user subscription at parent level
- Moved user data subscription logic from UserProfileInternal and UserProfileExternal to UserProfile component
- Updated UserProfileInternal to receive user data as a prop and refactored the component accordingly
- Updated UserProfileExternal to receive user data as a prop and refactored the component accordingly
- Ensured conditional rendering of UserProfileInternal or UserProfileExternal based on presence of userId in URL parameters
- Cleaned up child components by passing user data as props, thereby removing redundant useEffect hooks for subscriptions in both UserProfileInternal and UserProfileExternal

Changes:
1. **UserProfile.tsx**
   - Added useEffect to handle user data subscription.
   - Passed the subscribed user data as props to UserProfileInternal and UserProfileExternal.
   - Simplified the rendering logic to choose between internal and external profiles based on userId.

2. **UserProfileInternal.tsx**
   - Removed useEffect handling user data subscription.
   - Accepted  data as a prop.
   - Adjusted component logic to use the passed  prop for rendering user details.

3. **UserProfileExternal.tsx**
   - Removed useEffect handling user data subscription.
   - Accepted  data as a prop.
   - Adjusted component logic to use the passed  prop for rendering user details.

Why:
- Centralizing the user data subscription logic in the parent component () ensures a single source of truth for the user state.
- Improves code readability and maintainability by reducing redundancy.
- Ensures that user data is consistently managed and passed down to child components, avoiding potential issues with multiple subscriptions.
- Enhances separation of concerns, making child components more focused and easier to test.
